Refactor and support changing S3 Endpoint
See #470
