+++ .pre-commit-config.yaml
@@ -16,6 +16,16 @@ repos:
       - id: requirements-txt-fixer
       - id: check-added-large-files
       - id: check-merge-conflict
+  - repo: https://github.com/pycqa/isort
+    rev: 5.10.1
+    hooks:
+      - id: isort
+        name: isort (python)
+        args: [ "--profile", "black", "--filter-files" ]
+  - repo: https://github.com/psf/black
+    rev: 22.10.0
+    hooks:
+      - id: black
   - repo: https://github.com/pre-commit/mirrors-pylint
     rev: 'v3.0.0a5'  # Use the sha / tag you want to point at
     hooks:
@@ -24,7 +34,3 @@ repos:
     rev: '5.0.4'
     hooks:
       - id: flake8
-  - repo: https://github.com/psf/black
-    rev: 22.10.0
-    hooks:
-      - id: black
