+++ apps/cloud/odc/apps/cloud/s3_find.py
@@ -1,6 +1,7 @@
-import click
 import sys
-from odc.aio import s3_find_glob, S3Fetcher
+
+import click
+from odc.aio import S3Fetcher, s3_find_glob
 
 
 @click.command("s3-find")
@@ -56,7 +57,7 @@ def cli(uri, skip_check, no_sign_request=None, request_payer=False):
     except ValueError as ve:
         click.echo(str(ve), err=True)
         sys.exit(1)
-    except Exception as e:
+    except Exception as e: #pylint:disable=broad-except
         click.echo(str(e), err=True)
         sys.exit(1)
